requirements.txt
# PostgreSQL Database
psycopg2==2.8.3
# for testing auto deploy comment.

Django==2.2.1

django-hosts==3.0
# It is for WishWork.hosts file. (Needs settings configurations)

suds-jurko==0.6
zeep==3.4.0
# It is for zarrinPal's APICalls. (No need for configuration)

django-ckeditor==5.7.1 # It's for models RichTextField
django-js-asset==1.2.2 # It's dependent to djangoCKEditor
Pillow==6.0.0 # It's dependent to djangoCKEditor

djangorestframework==3.9.4

django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
# These two are related to each other.

django-cors-headers==3.0.1
# It is for outside request permission and security issues.

django-crontab==0.7.1

requests==2.22.0
# It is for being able to send grass-fed http1.1 sending requests. (API.Authentication.views)

pytz==2019.1
# It is for being able to timezone calculations using python. (API.Dashboard.serializers)

kavenegar==1.1.2
# SDK of kavenegar sms service

# It's for django asynchronous function calls
# In this package it should download Successfully installed celery pytz billiard kombu redis anyjson amqp
